requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
email-validator==2.0.0
pytest==8.3.5
pytest-playwright==0.5.1
playwright==1.47.0
